requirements.txt
arrow==0.16.0
asgiref==3.2.10
blessed==1.17.10
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
Django==3.1.1
django-appconf==1.0.4
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-cryptography==1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-picklefield==3.0.1
django-q==1.3.3
django-widget-tweaks==1.4.8
h11==0.9.0
httptools==0.1.1
idna==2.10
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
